deploy-package/ai/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
openai>=1.58.1
backoff>=2.2.1
numpy>=1.26.4
pandas>=2.2.3
scikit-learn>=1.5.2
pgvector==0.3.6
psycopg2-binary==2.9.10
pydantic==2.10.4
python-dotenv==1.0.1
httpx==0.28.1
tenacity==9.0.0
tiktoken>=0.8.0
redis==5.2.1
diskcache==5.6.3
langchain>=0.3.11
langchain-openai>=0.2.14

# Web tools and APIs
requests==2.31.0
aiohttp==3.8.5
beautifulsoup4==4.12.2

# Async support
asyncio==3.4.3

# Rate limiting and API management
ratelimit==2.2.1
python-jose==3.3.0

# State management and persistence
sqlitedict==2.1.0

# Voice API dependencies
google-cloud-texttospeech==2.14.1
google-cloud-speech==2.17.3
boto3==1.26.135
azure-cognitiveservices-speech==1.25.0
pydub==0.25.1
soundfile==0.12.1
librosa==0.10.0
ffmpeg-python==0.2.0

# OpenTelemetry dependencies
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-instrumentation-httpx==0.43b0
opentelemetry-instrumentation-psycopg2==0.43b0
opentelemetry-instrumentation-redis==0.43b0
opentelemetry-instrumentation-requests==0.43b0
opentelemetry-instrumentation-aiohttp-client==0.43b0
opentelemetry-instrumentation-sqlalchemy==0.43b0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-exporter-prometheus==0.43b0
prometheus-client==0.19.0
